third_party/WebKit/LayoutTests/inspector/console/console-format-es6-2-expected.txt

CONSOLE MESSAGE: line 11: [object Map Iterator]
CONSOLE MESSAGE: line 12: [object Map Iterator]
CONSOLE MESSAGE: line 11: [object Map Iterator]
CONSOLE MESSAGE: line 12: [object Map Iterator]
CONSOLE MESSAGE: line 11: [object Map Iterator]
CONSOLE MESSAGE: line 12: [object Map Iterator]
CONSOLE MESSAGE: line 11: [object Set Iterator]
CONSOLE MESSAGE: line 12: [object Set Iterator]
CONSOLE MESSAGE: line 11: [object Set Iterator]
CONSOLE MESSAGE: line 12: [object Set Iterator]
CONSOLE MESSAGE: line 11: [object Set Iterator]
CONSOLE MESSAGE: line 12: [object Set Iterator]
CONSOLE MESSAGE: line 11: [object Map Iterator]
CONSOLE MESSAGE: line 12: [object Map Iterator]
CONSOLE MESSAGE: line 11: [object Set Iterator]
CONSOLE MESSAGE: line 12: [object Set Iterator]
Tests that console properly displays information about ES6 features.

console-format-es6-2.html:11 MapIterator {41, Object {foo: 1}}
console-format-es6-2.html:12 [MapIterator]
globals[0]
MapIterator {41, Object {foo: 1}}
console-format-es6-2.html:11 MapIterator {42, Object {foo: 2}}
console-format-es6-2.html:12 [MapIterator]
globals[1]
MapIterator {42, Object {foo: 2}}
console-format-es6-2.html:11 MapIterator {41 => 42, Object {foo: 1} => Object {foo: 2}}
console-format-es6-2.html:12 [MapIterator]
globals[2]
MapIterator {41 => 42, Object {foo: 1} => Object {foo: 2}}
console-format-es6-2.html:11 SetIterator {41, Object {foo: 1}}
console-format-es6-2.html:12 [SetIterator]
globals[3]
SetIterator {41, Object {foo: 1}}
console-format-es6-2.html:11 SetIterator {41, Object {foo: 1}}
console-format-es6-2.html:12 [SetIterator]
globals[4]
SetIterator {41, Object {foo: 1}}
console-format-es6-2.html:11 SetIterator {41, Object {foo: 1}}
console-format-es6-2.html:12 [SetIterator]
globals[5]
SetIterator {41, Object {foo: 1}}
console-format-es6-2.html:11 MapIterator {Object {foo: 2}}
console-format-es6-2.html:12 [MapIterator]
globals[6]
MapIterator {Object {foo: 2}}
console-format-es6-2.html:11 SetIterator {Object {foo: 1}}
console-format-es6-2.html:12 [SetIterator]
globals[7]
SetIterator {Object {foo: 1}}
Expanded all messages
console-format-es6-2.html:11 MapIterator {41, Object {foo: 1}}
    __proto__: Map Iterator
    [[IteratorHasMore]]: true
    [[IteratorIndex]]: 0
    [[IteratorKind]]: "keys"
    [[Entries]]: Array(2)
        0: 41
        1: Object
        length: 2
console-format-es6-2.html:12 [MapIterator]
    0: MapIterator
    length: 1
    __proto__: Array(0)
globals[0]
MapIterator {41, Object {foo: 1}}
    __proto__: Map Iterator
    [[IteratorHasMore]]: true
    [[IteratorIndex]]: 0
    [[IteratorKind]]: "keys"
    [[Entries]]: Array(2)
        0: 41
        1: Object
        length: 2
console-format-es6-2.html:11 MapIterator {42, Object {foo: 2}}
    __proto__: Map Iterator
    [[IteratorHasMore]]: true
    [[IteratorIndex]]: 0
    [[IteratorKind]]: "values"
    [[Entries]]: Array(2)
        0: 42
        1: Object
        length: 2
console-format-es6-2.html:12 [MapIterator]
    0: MapIterator
    length: 1
    __proto__: Array(0)
globals[1]
MapIterator {42, Object {foo: 2}}
    __proto__: Map Iterator
    [[IteratorHasMore]]: true
    [[IteratorIndex]]: 0
    [[IteratorKind]]: "values"
    [[Entries]]: Array(2)
        0: 42
        1: Object
        length: 2
console-format-es6-2.html:11 MapIterator {41 => 42, Object {foo: 1} => Object {foo: 2}}
    __proto__: Map Iterator
    [[IteratorHasMore]]: true
    [[IteratorIndex]]: 0
    [[IteratorKind]]: "entries"
    [[Entries]]: Array(2)
        0: {41 => 42}
        1: {Object => Object}
        length: 2
console-format-es6-2.html:12 [MapIterator]
    0: MapIterator
    length: 1
    __proto__: Array(0)
globals[2]
MapIterator {41 => 42, Object {foo: 1} => Object {foo: 2}}
    __proto__: Map Iterator
    [[IteratorHasMore]]: true
    [[IteratorIndex]]: 0
    [[IteratorKind]]: "entries"
    [[Entries]]: Array(2)
        0: {41 => 42}
        1: {Object => Object}
        length: 2
console-format-es6-2.html:11 SetIterator {41, Object {foo: 1}}
    __proto__: Set Iterator
    [[IteratorHasMore]]: true
    [[IteratorIndex]]: 0
    [[IteratorKind]]: "values"
    [[Entries]]: Array(2)
        0: 41
        1: Object
        length: 2
console-format-es6-2.html:12 [SetIterator]
    0: SetIterator
    length: 1
    __proto__: Array(0)
globals[3]
SetIterator {41, Object {foo: 1}}
    __proto__: Set Iterator
    [[IteratorHasMore]]: true
    [[IteratorIndex]]: 0
    [[IteratorKind]]: "values"
    [[Entries]]: Array(2)
        0: 41
        1: Object
        length: 2
console-format-es6-2.html:11 SetIterator {41, Object {foo: 1}}
    __proto__: Set Iterator
    [[IteratorHasMore]]: true
    [[IteratorIndex]]: 0
    [[IteratorKind]]: "values"
    [[Entries]]: Array(2)
        0: 41
        1: Object
        length: 2
console-format-es6-2.html:12 [SetIterator]
    0: SetIterator
    length: 1
    __proto__: Array(0)
globals[4]
SetIterator {41, Object {foo: 1}}
    __proto__: Set Iterator
    [[IteratorHasMore]]: true
    [[IteratorIndex]]: 0
    [[IteratorKind]]: "values"
    [[Entries]]: Array(2)
        0: 41
        1: Object
        length: 2
console-format-es6-2.html:11 SetIterator {41, Object {foo: 1}}
    __proto__: Set Iterator
    [[IteratorHasMore]]: true
    [[IteratorIndex]]: 0
    [[IteratorKind]]: "entries"
    [[Entries]]: Array(2)
        0: 41
        1: Object
        length: 2
console-format-es6-2.html:12 [SetIterator]
    0: SetIterator
    length: 1
    __proto__: Array(0)
globals[5]
SetIterator {41, Object {foo: 1}}
    __proto__: Set Iterator
    [[IteratorHasMore]]: true
    [[IteratorIndex]]: 0
    [[IteratorKind]]: "entries"
    [[Entries]]: Array(2)
        0: 41
        1: Object
        length: 2
console-format-es6-2.html:11 MapIterator {Object {foo: 2}}
    __proto__: Map Iterator
    [[IteratorHasMore]]: true
    [[IteratorIndex]]: 1
    [[IteratorKind]]: "values"
    [[Entries]]: Array(1)
        0: Object
        length: 1
console-format-es6-2.html:12 [MapIterator]
    0: MapIterator
    length: 1
    __proto__: Array(0)
globals[6]
MapIterator {Object {foo: 2}}
    __proto__: Map Iterator
    [[IteratorHasMore]]: true
    [[IteratorIndex]]: 1
    [[IteratorKind]]: "values"
    [[Entries]]: Array(1)
        0: Object
        length: 1
console-format-es6-2.html:11 SetIterator {Object {foo: 1}}
    __proto__: Set Iterator
    [[IteratorHasMore]]: true
    [[IteratorIndex]]: 1
    [[IteratorKind]]: "values"
    [[Entries]]: Array(1)
        0: Object
        length: 1
console-format-es6-2.html:12 [SetIterator]
    0: SetIterator
    length: 1
    __proto__: Array(0)
globals[7]
SetIterator {Object {foo: 1}}
    __proto__: Set Iterator
    [[IteratorHasMore]]: true
    [[IteratorIndex]]: 1
    [[IteratorKind]]: "values"
    [[Entries]]: Array(1)
        0: Object
        length: 1
